fix: layered architecture completed

+++ src/services/FuncionarioService.java
@@ -5,10 +5,12 @@ import exceptions.AcessoNegadoException;
 import exceptions.FormatoIncorretoException;
 import exceptions.FuncionarioExistenteException;
 import model.Funcionario;
-import java.awt.*;
+
 import java.time.LocalDate;
 import java.util.List;
 import java.util.NoSuchElementException;
+
+import static io.Output.write;
 
 public class FuncionarioService {
     private final FuncionarioDAO funcionarioDAO;
@@ -17,20 +19,40 @@ public class FuncionarioService {
         this.funcionarioDAO = new FuncionarioDAO();
     }
 
-    public Integer cadastrarFuncionario(String nome, String cpf, String email,
+    public Integer cadastrarFuncionario(String nome, String cpf, String email, String senha,
                                             String cargo, LocalDate admissao) {
+        // — > Validações de entrada
+        if (!nome.matches("[A-Za-z]+\\s+[A-Za-z]{3,45}")) {
+            throw new FormatoIncorretoException("Nome invalido.");
+        }
 
         if (funcionarioDAO.findByCPF(cpf).isPresent()) {
             throw new FuncionarioExistenteException("Este CPF ja esta associado a um funcionario.");
         }
 
+        if (!cpf.matches("[\\d(-.)?]{11,25}")) {
+            throw new FormatoIncorretoException("CPF invalido.");
+        }
+        String newCpf = cpf.replaceAll("\\D", "");
+
+        if (!email.matches("[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,6}")) {
+            throw new FormatoIncorretoException("Email invalido.");
+        }
+
+        if (!senha.matches("\\d{6}")) {
+            throw new FormatoIncorretoException("Senha invalida.");
+        }
+
+        // — > Criação do objeto Funcionario para o envio ao banco
         Funcionario f = new Funcionario();
         f.setNome(nome);
-        f.setCpf(cpf);
+        f.setCpf(newCpf);
         f.setEmail(email);
+        f.setSenha(senha);
         f.setCargo(cargo);
-        f.setAdimissao(admissao);
+        f.setAdmissao(admissao);
 
+        // — > Insert do objeto Funcionario no banco de dados
         return funcionarioDAO.insert(f);
     }
 
@@ -38,49 +60,58 @@ public class FuncionarioService {
         if (!cpf.matches("[\\d(-.)?]{11,25}")) {
             throw new FormatoIncorretoException("CPF invalido.");
         }
-
         String newCpf = cpf.replaceAll("\\D", "");
 
         return funcionarioDAO.findByCPF(newCpf)
                 .orElseThrow(() -> new NoSuchElementException("CPF nao encontrado."));
     }
 
-    public Funcionario buscarFuncionario(Integer id) {
-        return funcionarioDAO.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("CPF nao encontrado."));
-    }
-
-    public void atualizarDados(String cpf, String nome, String email) {
-        if (cpf.matches("\\d{11,25}")) {
+    public void removerFuncionario(Funcionario funcionario, String cpf) {
+        if (!cpf.matches("[\\d(-.)?]{11,25}")) {
             throw new FormatoIncorretoException("CPF invalido.");
         }
 
-        Funcionario f = buscarFuncionario(cpf);
-        if (!nome.isEmpty()) f.setNome(nome);
-        if (!email.isEmpty()) f.setEmail(email);
-
-        funcionarioDAO.update(f);
-    }
-
-    public void realizarPromocao(String cpf, String cargo) {
-        if (cpf.matches("\\d{11,25}")) {
-            throw new FormatoIncorretoException("CPF invalido.");
+        if (cpf.equals(funcionario.getCpf())) {
+            throw new IllegalArgumentException("Nao e possivel fazer a remocao de si mesmo.");
         }
 
-        Funcionario f = buscarFuncionario(cpf);
+        buscarFuncionario(cpf);
+        funcionarioDAO.deleteByCPF(cpf);
+    }
+
+    public void atualizarDados(Funcionario funcionario, String nome, String email, String senha) {
+
+        if (nome != null) funcionario.setNome(nome);
+        if (email != null) funcionario.setEmail(email);
+        if (senha != null) funcionario.setSenha(senha);
+
+        funcionarioDAO.update(funcionario);
+    }
+
+    public void realizarPromocao(Funcionario funcionario, String cpf, String cargo) {
+        if (!cpf.matches("[\\d(-.)?]{11,25}")) {
+            throw new FormatoIncorretoException("CPF invalido.");
+        }
+        String newCpf = cpf.replaceAll("\\D", "");
+
+        if (cpf.equals(funcionario.getCpf())) {
+            throw new IllegalArgumentException("Nao e possivel fazer a promocao de si mesmo.");
+        }
+
+        Funcionario f = buscarFuncionario(newCpf);
         if (f.getCargo().equals(cargo)) throw new IllegalArgumentException("Este funcionario ja esta com este cargo atribuido.");
         f.setCargo(cargo);
 
         funcionarioDAO.update(f);
     }
 
-    public void removerFuncionario(String cpf) {
-        if (!cpf.matches("[\\d(-.)?]{11,25}")) {
-            throw new FormatoIncorretoException("CPF invalido.");
-        }
-
-        buscarFuncionario(cpf);
-        funcionarioDAO.deleteByCPF(cpf);
+    public List<Funcionario> buscaFiltradaFuncionario(String busca) {
+        return listarFuncionario()
+                .stream()
+                .filter(f -> f.getNome().equals(busca)
+                        || f.getCpf().equals(busca)
+                        || f.getEmail().equals(busca))
+                .toList();
     }
 
     public List<Funcionario> listarFuncionario() {
